telugu-agent-main/requirements.txt
# cache-bust: v3
livekit-agents>=0.12.0
livekit-plugins-google>=0.8.0
livekit-plugins-sarvam>=0.1.0
google-genai
python-dotenv
supabase
